// app/data/as568.generated.ts
export type As568Size = {
  dash: string;
  idIn: number;
  idToleranceIn: number;
  csIn: number;
  csToleranceIn: number;
};